lockingang/src/backend/rag_chatbot/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
pinecone>=4.1.0
python-dotenv>=1.0.0
pypdf>=4.2.0
python-docx>=1.1.0
python-multipart>=0.0.9
